+++ spec/acceptance/init_spec.rb
@@ -8,6 +8,10 @@ RSpec.describe 'phabricator' do
       package_manage => true,
     }
 
+    # Ensure that `apt-get update` is executed before any packages are
+    # installed. See https://github.com/puppetlabs/puppetlabs-apt/#adding-new-sources-or-ppas.
+    Class['apt::update'] -> Package <| title != 'apt-transport-https' and title != 'ca-certificates' and title != 'software-properties-common' |>
+
     class { 'php::globals':
       php_version => '7.1',
     }
@@ -33,6 +37,7 @@ RSpec.describe 'phabricator' do
     }
   EOS
 
+  # TODO: Move this to a shared example.
   it 'applies with no errors' do
     apply_manifest(pp, catch_failures: true)
   end